backend/requirements.txt
# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Image processing
opencv-python-headless>=4.9.0
pytesseract>=0.3.10
Pillow>=10.0.0

# HTTP client (for Scryfall API)
requests>=2.31.0

# Production server
gunicorn>=21.0.0

# Utilities
python-dotenv>=1.0.0
